AutoVulRepair-vulrag-branch/requirements.txt
Flask>=2.0,<3.0
Flask-Login>=0.6
Authlib>=1.2.0
requests>=2.25.0
python-dotenv>=1.0
celery>=5.2.0
redis>=4.0.0
sqlalchemy>=1.4.0
werkzeug>=2.0.0
docker>=6.0.0
pytest>=7.0.0
hypothesis>=6.0.0

# AI Patching System Dependencies
google-generativeai>=0.3.0
faiss-cpu>=1.7.4
sentence-transformers>=2.2.0
torch>=2.0.0
numpy>=1.21.0
tqdm>=4.65.0
